토마스 방향 버그 수정
타일 변경 지속 시간 자료형 변경 (int -> float)

+++ Assets/02. Scripts/Boss/TileSetData.cs
@@ -15,7 +15,7 @@ public class TileSetData
     public List<Vector2> tilePositions = new List<Vector2>();
     public TileType type;
     public float startupTime;
-    public int holdingTime;
+    public float holdingTime;
 
     [Header("기차 출현")]
     public bool useTrail;
